backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Youtube Quiz Generator Backend"
authors = ["Asako Hayase"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = "^0.1.24"
unstructured = '==0.10.25'
#Access weather data
pyowm = '3.3.0'
tools = "^0.1.9"
python-dotenv = "^1.0.1"

#For type error check
tiktoken = "^0.7.0"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
youtube-transcript-api = "^0.6.2"
groq = "^0.9.0"
[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

#Formatter
[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

#Declare that poetry is the package manager for this project
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


